LayoutTests/http/tests/xmlhttprequest/origin-allow-list-all-expected.txt

Tests the special case of allowlisting all origins.

Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (sync)
PASS: PASS
Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (async)
PASS: PASS
Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (sync)
PASS: PASS
Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (async)
PASS: PASS
